refactor(2024): ♻️ Update return types to int64 for better usage down the line.

+++ 2024/AoC2024/src/01/Day01.fs
@@ -29,7 +29,7 @@ let valueFolder (accumulator: int seq * int seq) (current: int * int) : int seq 
 
     (newLeft, newRight)
 
-let solve () : (int * int) =
+let solve () : (int64 * int64) =
     let (left, right) =
         Path.Combine(__SOURCE_DIRECTORY__, "input.txt")
         |> File.ReadAllLines
