+++ 2024/AoC2024/src/02/Day02.fs
@@ -37,7 +37,7 @@ let second (levels: int seq seq) : int =
     |> Seq.filter (fun value -> value = true)
     |> Seq.length
 
-let solve () : (int * int) =
+let solve () : (int64 * int64) =
     let levels =
         Path.Combine(__SOURCE_DIRECTORY__, "input.txt")
         |> File.ReadAllLines
